feat: Внедрение компонента "Выдвижная боковая панель" (Drawer) для CRUD-операций
Произведен полный рефакторинг UX для операций создания, редактирования и удаления карточек. Заменена механика модальных окон на новый, переиспользуемый компонент "Выдвижная боковая панель" (Drawer).

Ключевые изменения:
- Создан новый компонент `ui/drawer.tsx` с поддержкой анимаций (`framer-motion`), различных размеров и динамического контента.
- Внедрено централизованное управление состоянием панели с помощью `zustand` (`hooks/use-drawer.ts`).
- Логика формы была вынесена из модального окна в переиспользуемый компонент `admin/card-form.tsx`.
- Старый компонент `admin/card-form-modal.tsx` был удален.
- Созданы вспомогательные компоненты для содержимого панели: `drawer-card-actions.tsx` и `drawer-delete-confirmation.tsx`.
- Компонент `admin/cards-management.tsx` был полностью переписан для использования новой системы `Drawer`, реализуя все сценарии, описанные в ТЗ.
- Весь функционал был протестирован с помощью Playwright.
