+++ dev.log
@@ -1,5 +1,8 @@
 
-> my-v0-project@0.1.0 dev /app
+> my-v0-project@0.1.0 dev
 > next dev
 
- ⚠ Port 3000 is in use, trying 3001 instead.
+  ▲ Next.js 14.2.16
+  - Local:        http://localhost:3000
+
+ ✓ Starting...
